--- src/main/scala/Main.scala
@@ -1,9 +1,0 @@
-import palindrome.Palindrome
-import passwordChecker.PasswordChecker
-
-object Main {
-  def main(args: Array[String]): Unit = {
-    println(Palindrome.isPalindrome("mad"))
-    println(PasswordChecker.isValidPassword("b3stpAssw0rd!"))
-  }
-}
