+++ src/main/scala/algorithms/PasswordChecker.scala
@@ -1,4 +1,4 @@
-package passwordChecker
+package algorithms
 
 object PasswordChecker {
